Rename fileUploadSpecifications to be more accurate
